\name{modisRiceValidate}

\alias{modisRiceValidate}

\title{ Validates rice from the previously generated rice pixels by using Enhanced Vegetation Index(EVI) values. }

\description{
This function reads the rice pixels generated from Vegetation indices. Then, it validates the rice areas by using
Enhanced Vegetation Indices (EVI) calculated per 8-day composite from the previous and current year.
It identifies rice pixels by checking the rice temporal pattern using EVI values.
This function uses backtracking algorithm which tracks the maximum EVI value, the half of the maximum EVI,
and EVI values less than the 0.2 threshold for flooded (Xiao,2006).
}

\usage{
modisRiceValidate(perhapsPath, curYearPath, prevYearPath, outPath, tileNumber)
}

\arguments{
  \item{perhapsPath}{ path with perhapsRice file (possible rice areas). }
  \item{curYearPath}{ path with evi tiff files for the current year of the possible rice areas. }
  \item{prevYearPath}{ path with evi tiff files for the previous year of the possible rice areas. }
  \item{outPath}{ path which the verified rice areas will be stored }
  \item{tileNumber}{ vertical and horizontal tile number of MODIS, tileNumber should follow the format in the MODIS filename, e.g. "h29v07"  }
  }

\details{
  Uses EVI images that has been generated by modisVeg as input in verifying MODIS rice areas.
  }

\author{ Angelo Carlo Pacheco }

\seealso{ \code{\link{modisClean}}, \code{\link{modisVeg}}}

\note{ 

# to use this function, follow these steps:

  setwd(dirname(system.file("external/.", package="RemoteSensing")))

# set the perhapsRice directory path
# the path varies depending on your Operating System
# this one is for Linux system:
  perhapsPath <- "/home/yann/Desktop/evi/rice"
  
# In Windows OS: perhapsPath <- "C:\\path\\of\\the\\directory\\for\\perhapsRice"
# Note: replace '\' by '\\' or '/' so R will be able to read it as directory path

# set directory path of evi tiff files for the current year of the possible rice areas
  curYearPath <- "/home/yann/Desktop/evi/2007"

# set directory path of evi tiff files for the previous year of the possible rice areas
  prevYearPath <- "/home/yann/Desktop/evi/2006"

# set directory path of the folder where the output will be saved
  outPath <- "/home/yann/Desktop/evi/reallyRice"

# set the tile number
  tileNumber <- "h29v07"

# call this function
  modisRiceValidate(perhapsPath, curYearPath, prevYearPath, outPath, tileNumber)

 }
\references{Xiao, X., Boles, S., Frolking, S. et.al. 2006. Mapping paddy rice agriculture in Southeast Asia using multi-temporal MODIS images.  Remote Sensing of Environment 100, pages, 95-113.}

