\name{RemoteSensing-package}
\alias{RemoteSensing-package}
\alias{RemoteSensing}
\docType{package}
\title{
What the package does (short line)
Remote sensing
}
\description{
Analysis of satellite images. 

}
\details{
\tabular{ll}{
Package: \tab RemoteSensing\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2008-10-09\cr
License: \tab GPL v3\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Yann Chemin

Maintainer: Who to complain to <y.chemin@gmail.com>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\keyword{ spatial }

\seealso{
\code{\link[<raster>:<raster>-package]{<raster>}}
}
\examples{
}
