\name{Snow indiex}

\alias{ndsi}

\title{Snow index}

\description{
  Compute vegetation indices from remote sensing data. 
}

\usage{
ndsi(green, nir)
}


\arguments{
  \item{green}{green channel (band) data}
  \item{nir}{nir channel (band) data}
}

\value{
  snow index
}

\author{Yann Chemin}


\keyword{spatial}

