\name{modis}

\alias{modisFiles}


\title{ modis files }

\description{
Get a table with parsed Modis files
}

\usage{
modisFiles(path)
}


\arguments{
  \item{path}{ path (folder) where the files are}
}
  
\value{
data.frame
}

\author{ Robert Hijmans and Sonia Asilo }

\keyword{ spatial }

